requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1
pandas>=2.0
openpyxl>=3.1
Pillow>=10.0
reportlab>=4.0
python-barcode>=0.15
djangorestframework>=3.15,<4.0
django-filter>=24.0
requests>=2.31

gunicorn>=21.2
whitenoise>=6.6
cryptography>=41.0
